requirements.txt
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.5
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
Scikit-learn==0.23.2
